Reorganized routes.js into multiple route files
Extra fixes:
- Moved socket.io bits into home.js route
- Ensure that client side always returns some sort of coordinates
