+++ controllers/signup.js
@@ -32,21 +32,19 @@ var router = express.Router();
 
 module.exports = function(sio, passport) {
 
-    // other routes
-    router.use('/login', require('./login')(sio, passport));
-    router.use('/logout', require('./logout')(sio, passport));
-    router.use('/home', require('./home')(sio, passport));
-    router.use('/profile', require('./profile')(sio, passport));
-    router.use('/user-prefs', require('./user-prefs')(sio, passport));
-    router.use('/meeting', require('./meeting')(sio, passport));
-    router.use('/results', require('./results')(sio, passport));
-    router.use('/create', require('./create')(sio, passport));
-    router.use('/signup', require('./signup')(sio, passport));
+    router.get('/', function (req, res) {
+        // render the page and pass in any flash data if it exists
+        res.render('signup.ejs', {
+            message: req.flash('signupMessage')
+        });
+    });
 
-    //Home page (with login links)
-	router.get('/', function (req, res) {
-		res.render('index');
-	});
+    //process the signup form
+    router.post('/', passport.authenticate('local-signup', {
+        successRedirect: '/home', // redirect to the secure home section
+        failureRedirect: '/signup', // redirect back to the signup page if there is an error
+        failureFlash: true // allow flash messages
+    }));
 
     return router;
 }
