+++ npm-debug.log
@@ -1,0 +1,47 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/home/noah/.nvm/versions/node/v6.4.0/bin/node',
+1 verbose cli   '/home/noah/.nvm/versions/node/v6.4.0/bin/npm',
+1 verbose cli   'start' ]
+2 info using npm@3.10.3
+3 info using node@v6.4.0
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info lifecycle meeting-magic@~prestart: meeting-magic@
+6 silly lifecycle meeting-magic@~prestart: no script for prestart, continuing
+7 info lifecycle meeting-magic@~start: meeting-magic@
+8 verbose lifecycle meeting-magic@~start: unsafe-perm in lifecycle true
+9 verbose lifecycle meeting-magic@~start: PATH: /home/noah/.nvm/versions/node/v6.4.0/lib/node_modules/npm/bin/node-gyp-bin:/home/noah/Meeting-Magic/node_modules/.bin:/home/noah/.nvm/versions/node/v6.4.0/bin:/home/noah/.nvm/versions/node/v6.4.0/bin:/home/noah/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin
+10 verbose lifecycle meeting-magic@~start: CWD: /home/noah/Meeting-Magic
+11 silly lifecycle meeting-magic@~start: Args: [ '-c', 'node server.js' ]
+12 silly lifecycle meeting-magic@~start: Returned: code: 1  signal: null
+13 info lifecycle meeting-magic@~start: Failed to exec start script
+14 verbose stack Error: meeting-magic@ start: `node server.js`
+14 verbose stack Exit status 1
+14 verbose stack     at EventEmitter.<anonymous> (/home/noah/.nvm/versions/node/v6.4.0/lib/node_modules/npm/lib/utils/lifecycle.js:242:16)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at EventEmitter.emit (events.js:191:7)
+14 verbose stack     at ChildProcess.<anonymous> (/home/noah/.nvm/versions/node/v6.4.0/lib/node_modules/npm/lib/utils/spawn.js:40:14)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at ChildProcess.emit (events.js:191:7)
+14 verbose stack     at maybeClose (internal/child_process.js:852:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:215:5)
+15 verbose pkgid meeting-magic@
+16 verbose cwd /home/noah/Meeting-Magic
+17 error Linux 4.4.0-34-generic
+18 error argv "/home/noah/.nvm/versions/node/v6.4.0/bin/node" "/home/noah/.nvm/versions/node/v6.4.0/bin/npm" "start"
+19 error node v6.4.0
+20 error npm  v3.10.3
+21 error code ELIFECYCLE
+22 error meeting-magic@ start: `node server.js`
+22 error Exit status 1
+23 error Failed at the meeting-magic@ start script 'node server.js'.
+23 error Make sure you have the latest version of node.js and npm installed.
+23 error If you do, this is most likely a problem with the meeting-magic package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     node server.js
+23 error You can get information on how to open an issue for this project with:
+23 error     npm bugs meeting-magic
+23 error Or if that isn't available, you can get their info via:
+23 error     npm owner ls meeting-magic
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
