+++ views/pages/meeting.ejs
@@ -60,7 +60,21 @@
 		function fillInAddress() {
 			// Get the place details from the autocomplete object.
 			var place = autocomplete.getPlace();
-			document.getElementById('location').value = place.geometry.location.lat() + "," + place.geometry.location.lng();
+			var lat;
+			var long;
+
+			if (place.geometry.location.lat()) {
+				lat = place.geometry.location.lat();
+			} else {
+				lat = navigator.geolocation.getCurrentPosition.position.coords.latitude;
+			}
+
+			if (place.geometry.location.lng()) {
+				long = place.geometry.location.lng();
+			} else {
+				long = navigator.geolocation.getCurrentPosition.position.coords.longitude;
+			}
+			document.getElementById('location').value = lat + "," + long;
 			console.log(document.getElementById('location').value);
 			/*for (var component in componentForm) {
 				document.getElementById(component).value = '';
